travel_agent_pro/requirements.txt
# ── Core Web Framework ──────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── HTTP Client ──────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── AI / LLM ─────────────────────────────────────────────────────────────────
openai>=1.30.0

# ── Notion Integration ───────────────────────────────────────────────────────
notion-client>=2.2.1

# ── Web Search ───────────────────────────────────────────────────────────────
tavily-python>=0.3.3

# ── Environment Variables ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Data Validation ──────────────────────────────────────────────────────────
pydantic>=2.0.0

# ── Frontend (Streamlit UI) ───────────────────────────────────────────────────
streamlit>=1.35.0
extra-streamlit-components>=0.1.60
